<style> : Follow button 파일구조 변경
- 파일구조변경
