+++ src/components/common/Button/Button.jsx
@@ -1,7 +1,16 @@
-import { StyledBtn, SIZES } from './StyledBtn';
+import { StyledBtn, SIZES, COLOR } from './StyledBtn';
 
-const Button = ({ type, name, size, disabled, className, onClick }) => {
+const Button = ({
+  type,
+  name,
+  size,
+  disabled,
+  className,
+  onClick,
+  btncolor,
+}) => {
   const sizeStyle = SIZES[size];
+  const btnColor = COLOR[btncolor];
   return (
     <StyledBtn
       type={type}
@@ -9,6 +18,7 @@ const Button = ({ type, name, size, disabled, className, onClick }) => {
       sizeStyle={sizeStyle}
       disabled={disabled}
       onClick={onClick}
+      btncolor={btnColor}
     >
       {name}
     </StyledBtn>
